se crea la capa de controladores y sus respectivas hojas de controladores a cada modelo , se crean los controladores y se comprueba que se puedan guardar datos y listarlos en thunder de manera efectiva se envian un post y un get por cada modelo

+++ src/main/java/com/example/Gastapp/repositorios/ICategoriaRepositorio.java
@@ -6,11 +6,8 @@ import org.springframework.stereotype.Repository;
 import com.example.Gastapp.modelos.Categoria;
 
 @Repository
-public interface ICategoriaRepositorio extends JpaRepository<Categoria,Integer> {
+public interface ICategoriaRepositorio extends JpaRepository<Categoria, Long> {
 
-    //Espacio para crear consultas personalizadas
-    
-
-    
+    // Espacio para consultas personalizadas
 
 }
