+++ src/main/java/com/example/Gastapp/repositorios/IGastoRepositorio.java
@@ -6,11 +6,8 @@ import org.springframework.stereotype.Repository;
 import com.example.Gastapp.modelos.Gasto;
 
 @Repository
-public interface IGastoRepositorio extends JpaRepository<Gasto,Integer> {
+public interface IGastoRepositorio extends JpaRepository<Gasto, Long> {
 
-    //Espacio para crear consultas personalizadas
-    
-
-    
+    // Espacio para consultas personalizadas
 
 }
